+++ NEWS.md
@@ -3,3 +3,4 @@
 * Added a `NEWS.md` file to track changes to the package.
 * Use `readr::read_table()` instead of `readr::read_delim()` in `aoc_input_data_frame()`
 * Add user agent to `aoc_get_input()`
+* Add `extract_numbers()` to extract all numbers from a string